backend/requirements.txt
absl-py==2.1.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
bleach==6.1.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.2.2
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
chex==0.1.87
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cookiecutter==2.5.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Django==5.0.1
django-browser-reload==1.12.1
django-cors-headers==4.3.1
django-tailwind==3.8.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
etils==1.10.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.3.25
flax==0.10.0
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.10.0
gast==0.6.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.2
humanize==4.11.0
idna==3.6
imageio==2.35.1
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.34
jaxlib==0.4.34
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.7.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mpi4py==3.1.5
msgpack==1.1.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
optax==0.2.3
optree==0.12.1
orbax-checkpoint==0.8.0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.17.2
pynput==1.7.7
pyparsing==3.1.2
python-dateutil==2.8.2
python-json-logger==2.0.7
python-slugify==8.0.3
pytz==2023.4
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.2.0
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.20.0
rsa==4.9
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.1
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.4.4
stack-data==0.6.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow-decision-forests==1.8.1
tensorflow-estimator==2.15.0
tensorflow-hub==0.16.1
tensorflow-intel==2.15.1
tensorflow-io-gcs-filesystem==0.31.0
tensorflowjs==4.22.0
tensorstore==0.1.67
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
tf_keras==2.15.1
threadpoolctl==3.5.0
tifffile==2024.8.10
tinycss2==1.3.0
tomli==2.0.1
toolz==1.0.0
tornado==6.4.1
traitlets==5.14.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
uri-template==1.3.0
urllib3==2.2.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.14.1
wurlitzer==3.1.1
zipp==3.20.2
